requirements.txt
torch
torchvision
numpy
pyyaml
tqdm
pillow
matplotlib
tabulate
lpips
thop
tensorboard
